deploy/requirements.txt
uvicorn
torch
fastapi
streamlit
requests
torchvision
pillow
python-multipart